- upgrade chromedriver to 77 - move utils to popup-utils folder - fix Vladi's initial comments
