+++ packages/main/src/Popover.hbs
@@ -13,7 +13,7 @@
 	<span class="first-fe" tabindex="0" @focusin={{forwardToLast}}></span>	
 
 	<div class="ui5-popover-content">
-			<slot></slot>
+		<slot></slot>
 	</div>
 
 	<span class="last-fe" tabindex="0" @focusin={{forwardToFirst}}></span>
