+++ packages/main/src/popup-utils/PopupUtils.js
@@ -32,7 +32,7 @@ const isClickInRect = (event, rect) => {
 const getClosedPopupParent = el => {
 	const parent = el.parentElement || (el.getRootNode && el.getRootNode().host);
 
-	if (parent.openBy || parent === document.documentElement) {
+	if ((parent.openBy && parent.isUI5Element) || parent === document.documentElement) {
 		return parent;
 	}
